Research Assistant Agent/requirements.txt
crewai
python-dotenv
requests
beautifulsoup4
streamlit